Add borrowings end-point for the API.
